Move non-standardized functions (nan functions, pad, map_overlap) to separate package
