+++ cubed/tests/array/test_nan_functions.py
