+++ cubed/tests/array/test_pad.py
